sousearch/requirements.txt
Django==1.8.4
django-bootstrap-pagination==1.5.1
django-haystack==2.4.0
elasticsearch==1.6.0
MySQL-python==1.2.5
requests==2.7.0
urllib3==1.11
wheel==0.24.0
